backend/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.0
pymongo==4.6.2
PyJWT==2.8.0
python-dotenv==1.0.1
google-genai==1.2.0
bcrypt==4.1.2
Werkzeug==3.0.1
redis>=5.0.0
